requirements.txt
streamlit==1.42.0
google-generativeai
python-dotenv
